maint: Check for unnecessary casts
Casting an expression to the type that it already is, is flagged by the
clang-tidy check 'readability-redundant-casting'.

+++ gi/value.cpp
@@ -991,7 +991,7 @@ static JS::Value convert_int_to_enum(const GI::Repository& repo, GType gtype,
         // Native enums don't have type info, assume they are signed to avoid
         // crashing when they are exposed to JS.
         if (!info) {
-            v_double = static_cast<int64_t>(v);
+            v_double = v;
         } else {
             v_double = info->enum_from_int(v);
         }
